fix: export department management router
