+++ app.js
@@ -85,7 +85,7 @@ const stitchingPaymentRoutes = require('./routes/stitchingPaymentRoutes');
 const inventoryRoutes = require('./routes/inventoryRoutes');
 const departmentMgmtRoutes = require('./routes/departmentMgmtRoutes');
 const employeeRoutes = require('./routes/employeeRoutes');
-const salaryRoutes = require('./routes/salaryRoutes');
+const dihariRoutes = require('./routes/dihadiRoutes');
 const operatorEmployeeRoutes = require('./routes/operatorEmployeeRoutes');
 const flipkartReturnRoutes = require('./routes/flipkartReturnRoutes');
 const flipkartIssueStatusRoutes = require('./routes/flipkartIssueStatusRoutes');
@@ -117,7 +117,7 @@ app.use('/catalogupload', catalogR);
 app.use('/inventory', inventoryRoutes);
 app.use('/store-admin', storeAdminRoutes);
 app.use('/supervisor', employeeRoutes);
-app.use('/', salaryRoutes);
+app.use('/', dihariRoutes);
 app.use('/flipkart', flipkartReturnRoutes);
 app.use('/flipkart', flipkartIssueStatusRoutes);
 app.use('/webhook', inventoryWebhook);
