+++ helpers/salaryCalculator.js
@@ -1,10 +1,5 @@
 const moment = require('moment');
-const { SPECIAL_DEPARTMENTS } = require('../utils/departments');
-const {
-  SPECIAL_SUNDAY_SUPERVISORS,
-  FULL_SALARY_EMPLOYEE_IDS
-} = require('../utils/supervisors');
-const { HOURLY_EXEMPT_EMPLOYEE_IDS } = require('../utils/hourlyExemptEmployees');
+// Removed monthly salary utilities; only moment is required for dihadi calculations.
 
 function lunchDeduction(
   punchIn,
@@ -46,14 +41,6 @@ function effectiveHours(punchIn, punchOut, salaryType = 'dihadi', allottedHours 
   const start = moment(punchIn, 'HH:mm:ss');
   const end = moment(punchOut, 'HH:mm:ss');
   let mins = end.diff(start, 'minutes');
-  // Apply 15 minute grace period for monthly employees
-  if (salaryType !== 'dihadi') {
-    const shiftStart = moment('09:00:00', 'HH:mm:ss');
-    if (start.isAfter(shiftStart)) {
-      const grace = Math.min(15, start.diff(shiftStart, 'minutes'));
-      mins += grace;
-    }
-  }
   mins -= lunchDeduction(punchIn, punchOut, salaryType, allottedHours);
 
   // Deduct for late arrival after the 15 minute grace period
@@ -82,265 +69,11 @@ exports.effectiveHours = effectiveHours;
 
 async function calculateSalaryForMonth(conn, employeeId, month) {
   const [[emp]] = await conn.query(
-    `SELECT e.salary, e.salary_type, e.paid_sunday_allowance, e.pay_sunday, e.allotted_hours,
-            e.date_of_joining, e.designation,
-            d.name AS department,
-            u.username AS supervisor_name
-       FROM employees e
-       JOIN users u ON e.supervisor_id = u.id
-       LEFT JOIN (
-             SELECT user_id, MIN(department_id) AS department_id
-               FROM department_supervisors
-              GROUP BY user_id
-       ) ds ON ds.user_id = e.supervisor_id
-       LEFT JOIN departments d ON ds.department_id = d.id
-      WHERE e.id = ?`,
+    'SELECT salary, salary_type, allotted_hours, date_of_joining FROM employees WHERE id = ?',
     [employeeId]
   );
-  if (!emp) return;
-  const specialDept = SPECIAL_DEPARTMENTS.includes(
-    (emp.department || '').toLowerCase()
-  );
-
-  // Check designation. Salary for checkers is handled manually.
-  if ((emp.designation || '').toLowerCase() === 'checker') {
-    await conn.query(
-      'DELETE FROM employee_salaries WHERE employee_id = ? AND month = ?',
-      [employeeId, month]
-    );
-    return;
-  }
-  if (FULL_SALARY_EMPLOYEE_IDS.includes(employeeId)) {
-    const gross = parseFloat(emp.salary);
-    await conn.query(
-      `INSERT INTO employee_salaries (employee_id, month, gross, deduction, net, created_at)
-       VALUES (?, ?, ?, 0, ?, NOW())
-       ON DUPLICATE KEY UPDATE gross=VALUES(gross), deduction=0, net=VALUES(net)`,
-      [employeeId, month, gross, gross]
-    );
-    return;
-  }
-  const specialSup = SPECIAL_SUNDAY_SUPERVISORS.map(s => s.toLowerCase()).includes(
-    (emp.supervisor_name || '').toLowerCase()
-  );
-  if (emp.salary_type === 'dihadi') {
-    await calculateDihadiMonthly(conn, employeeId, month, emp);
-    return;
-  }
-  const monthStart = moment(month + '-01');
-  const join = emp.date_of_joining ? moment(emp.date_of_joining) : null;
-  const start = join && join.isAfter(monthStart) ? join.format('YYYY-MM-DD') : monthStart.format('YYYY-MM-DD');
-
-  const [attendance] = await conn.query(
-    'SELECT date, status, punch_in, punch_out FROM employee_attendance WHERE employee_id = ? AND date >= ? AND DATE_FORMAT(date, "%Y-%m") = ?',
-    [employeeId, start, month]
-  );
-  const [sandwichRows] = await conn.query(
-    'SELECT date FROM sandwich_dates WHERE DATE_FORMAT(date, "%Y-%m") = ?',
-    [month]
-  );
-  const daysInMonth = moment(month + '-01').daysInMonth();
-  const dailyRate = parseFloat(emp.salary) / daysInMonth;
-  const sandwichDates = sandwichRows.map(r => moment(r.date).format('YYYY-MM-DD'));
-  const attMap = {};
-  attendance.forEach(a => {
-    attMap[moment(a.date).format('YYYY-MM-DD')] = a.status;
-  });
-
-
-
-  // Treat missing attendance records as absences
-  const begin = join && join.isAfter(monthStart) ? join.date() : 1;
-  for (let d = begin; d <= daysInMonth; d++) {
-    const dateStr = moment(`${month}-${String(d).padStart(2, '0')}`).format('YYYY-MM-DD');
-    if (!attMap[dateStr]) {
-      attendance.push({ date: dateStr, status: 'absent', punch_in: null, punch_out: null });
-      attMap[dateStr] = 'absent';
-    }
-  }
-
-  // Ensure attendance is processed chronologically so Sunday rules apply
-  attendance.sort((a, b) => moment(a.date).diff(moment(b.date)));
-
-
-  // If an employee works on Sunday but misses Saturday or Monday,
-  // the adjacent absence should be paid. Collect those dates here
-  const skipAbsent = new Set();
-  if (!specialSup) {
-    attendance.forEach(a => {
-      if (moment(a.date).day() !== 0) return;
-      if (a.status !== 'present') return;
-      const satKey = moment(a.date).subtract(1, 'day').format('YYYY-MM-DD');
-      const monKey = moment(a.date).add(1, 'day').format('YYYY-MM-DD');
-      const satStatus = attMap[satKey];
-      const monStatus = attMap[monKey];
-      if (satStatus && (satStatus === 'absent' || satStatus === 'one punch only')) {
-        skipAbsent.add(satKey);
-      }
-      if (monStatus && (monStatus === 'absent' || monStatus === 'one punch only')) {
-        skipAbsent.add(monKey);
-      }
-    });
-  }
-
-  let absent = 0;
-  let halfDeduct = 0;
-  let extraPay = 0;
-  let mandatoryUsed = 0;
-  const creditLeaves = [];
-
-  attendance.forEach(a => {
-    const dateStr = moment(a.date).format('YYYY-MM-DD');
-    if (skipAbsent.has(dateStr)) {
-      return;
-    }
-    const status = a.status;
-    const isSun = moment(a.date).day() === 0;
-    const isSandwich = !specialSup && sandwichDates.includes(dateStr);
-
-    if (!specialSup && isSun) {
-      const satKey = moment(a.date).subtract(1, 'day').format('YYYY-MM-DD');
-      const monKey = moment(a.date).add(1, 'day').format('YYYY-MM-DD');
-      const satStatus = attMap[satKey];
-      const monStatus = attMap[monKey];
-      const missedSat = satStatus === 'absent' || satStatus === 'one punch only';
-      const missedMon = monStatus === 'absent' || monStatus === 'one punch only';
-      if (status === 'present') {
-        if (satStatus && missedSat) skipAbsent.add(satKey);
-        if (monStatus && missedMon) skipAbsent.add(monKey);
-      } else if (status !== 'present' && (missedSat || missedMon)) {
-        absent++;
-        return;
-      }
-    } else if (specialSup && isSun) {
-      const satKey = moment(a.date).subtract(1, 'day').format('YYYY-MM-DD');
-      const monKey = moment(a.date).add(1, 'day').format('YYYY-MM-DD');
-      const satStatus = attMap[satKey];
-      const monStatus = attMap[monKey];
-      const missedSat = satStatus === 'absent' || satStatus === 'one punch only';
-      const missedMon = monStatus === 'absent' || monStatus === 'one punch only';
-      if (status !== 'present' && missedSat && missedMon) {
-        absent++;
-        return;
-      }
-    }
-
-    if (isSandwich) {
-      const prevStatus = attMap[moment(a.date).subtract(1, 'day').format('YYYY-MM-DD')];
-      const nextStatus = attMap[moment(a.date).add(1, 'day').format('YYYY-MM-DD')];
-      const adjAbsent = (prevStatus === 'absent' || prevStatus === 'one punch only') ||
-                        (nextStatus === 'absent' || nextStatus === 'one punch only');
-      if (adjAbsent) {
-        absent++;
-        return;
-      }
-    }
-
-    if (isSun) {
-      const satKey = moment(a.date).subtract(1, 'day').format('YYYY-MM-DD');
-      const monKey = moment(a.date).add(1, 'day').format('YYYY-MM-DD');
-      const satStatus = attMap[satKey];
-      const monStatus = attMap[monKey];
-      const missedSat = satStatus === 'absent' || satStatus === 'one punch only';
-      const missedMon = monStatus === 'absent' || monStatus === 'one punch only';
-
-      if (!specialSup && status !== 'present' && (missedSat || missedMon)) {
-        absent++;
-        return;
-      }
-
-      if (!specialSup && !specialDept && status !== 'present' && mandatoryUsed < (emp.paid_sunday_allowance || 0)) {
-        mandatoryUsed++;
-        absent++;
-        return;
-      }
-
-      if (status === 'present' && a.punch_in && a.punch_out) {
-        const hrsWorked = effectiveHours(a.punch_in, a.punch_out, 'monthly');
-        if (hrsWorked > 0) {
-          let sundayPaid = false;
-          if (specialSup) {
-            extraPay += dailyRate;
-            sundayPaid = true;
-          } else if (specialDept) {
-            if (!skipAbsent.has(satKey) && !skipAbsent.has(monKey)) {
-              creditLeaves.push(dateStr);
-            }
-          } else if (mandatoryUsed < (emp.paid_sunday_allowance || 0)) {
-            mandatoryUsed++;
-          } else if (emp.pay_sunday) {
-            extraPay += dailyRate;
-            sundayPaid = true;
-          } else if (!skipAbsent.has(satKey) && !skipAbsent.has(monKey)) {
-            creditLeaves.push(dateStr);
-          }
-
-          if (sundayPaid) {
-            skipAbsent.delete(monKey);
-          }
-        }
-      }
-    } else {
-      if (status === 'absent' || status === 'one punch only' || !a.punch_in || !a.punch_out) {
-        absent++;
-      } else if (emp.allotted_hours) {
-        const hrsWorked = effectiveHours(a.punch_in, a.punch_out, 'monthly');
-        const allotted = parseFloat(emp.allotted_hours);
-        if (hrsWorked < allotted * 0.4) {
-          absent++;
-        } else if (hrsWorked < allotted * 0.85) {
-          halfDeduct += 0.5;
-        }
-      }
-    }
-  });
-
-  if (!specialSup) {
-    for (const d of creditLeaves) {
-      const [rows] = await conn.query(
-        'SELECT id FROM employee_leaves WHERE employee_id = ? AND leave_date = ?',
-        [employeeId, d]
-      );
-      if (!rows.length) {
-        await conn.query(
-          'INSERT INTO employee_leaves (employee_id, leave_date, days, remark) VALUES (?, ?, 1, ?)',
-          [employeeId, d, 'Sunday Credit']
-        );
-      }
-    }
-  }
-
-  const [leaveRows] = await conn.query(
-    `SELECT COALESCE(SUM(days),0) AS used
-       FROM employee_leaves
-      WHERE employee_id = ?
-        AND DATE_FORMAT(leave_date, "%Y-%m") = ?
-        AND (remark IS NULL OR LOWER(remark) <> 'sunday credit')`,
-    [employeeId, month]
-  );
-  const leavesUsed = parseFloat(leaveRows[0].used) || 0;
-  absent = Math.max(0, absent - leavesUsed);
-
-  const [nightRows] = await conn.query(
-    'SELECT COALESCE(SUM(nights),0) AS total_nights FROM employee_nights WHERE employee_id = ? AND month = ?',
-    [employeeId, month]
-  );
-  const nightPay = (parseFloat(nightRows[0].total_nights) || 0) * dailyRate;
-  extraPay += nightPay;
-  const gross = parseFloat(emp.salary) + extraPay;
-  const [[advRow]] = await conn.query(
-    'SELECT COALESCE(SUM(amount),0) AS total FROM advance_deductions WHERE employee_id = ? AND month = ?',
-    [employeeId, month]
-  );
-  const advDeduct = parseFloat(advRow.total) || 0;
-  const deduction = (absent + halfDeduct) * dailyRate + advDeduct;
-  const net = gross - deduction;
-  await conn.query(
-    `INSERT INTO employee_salaries (employee_id, month, gross, deduction, net, created_at)
-     VALUES (?, ?, ?, ?, ?, NOW())
-     ON DUPLICATE KEY UPDATE gross=VALUES(gross), deduction=VALUES(deduction), net=VALUES(net)`,
-    [employeeId, month, gross, deduction, net]
-  );
+  if (!emp || emp.salary_type !== 'dihadi') return;
+  await calculateDihadiMonthly(conn, employeeId, month, emp);
 }
 
 // Dihadi workers are paid purely on hours worked. Grab all attendance
@@ -373,92 +106,3 @@ async function calculateDihadiMonthly(conn, employeeId, month, emp) {
 }
 
 exports.calculateSalaryForMonth = calculateSalaryForMonth;
-
-// Hourly salary calculation for monthly employees
-async function calculateHourlyMonthly(conn, employeeId, month, emp, sundayHoursOverride = null) {
-  const monthStart = moment(month + '-01');
-  const join = emp.date_of_joining ? moment(emp.date_of_joining) : null;
-  const start = join && join.isAfter(monthStart) ? join.format('YYYY-MM-DD') : monthStart.format('YYYY-MM-DD');
-  const [attendance] = await conn.query(
-    'SELECT date, status, punch_in, punch_out FROM employee_attendance WHERE employee_id = ? AND date >= ? AND DATE_FORMAT(date, "%Y-%m") = ? ORDER BY date',
-    [employeeId, start, month]
-  );
-  const attMap = {};
-  attendance.forEach(a => {
-    attMap[moment(a.date).format('YYYY-MM-DD')] = a;
-  });
-  const monthStartHM = moment(month + '-01');
-  const startHM = join && join.isAfter(monthStartHM) ? join.date() : 1;
-  const daysInMonth = monthStartHM.daysInMonth();
-  const dailyRate = parseFloat(emp.salary) / daysInMonth;
-  const hourlyRate = emp.allotted_hours ? dailyRate / parseFloat(emp.allotted_hours) : 0;
-  const sundayBaseHours = sundayHoursOverride || 9;
-  const sundayRate = dailyRate / sundayBaseHours;
-  let totalPay = 0;
-  const graceTime = moment('09:15:00', 'HH:mm:ss');
-  for (let d = startHM; d <= daysInMonth; d++) {
-    const dateStr = moment(month + '-' + String(d).padStart(2, '0')).format('YYYY-MM-DD');
-    const rec = attMap[dateStr];
-    const isSun = moment(dateStr).day() === 0;
-    if (isSun) {
-      const prevKey = moment(dateStr).subtract(1, 'day').format('YYYY-MM-DD');
-      const nextKey = moment(dateStr).add(1, 'day').format('YYYY-MM-DD');
-      const prevStatus = attMap[prevKey] ? attMap[prevKey].status : null;
-      const nextStatus = attMap[nextKey] ? attMap[nextKey].status : null;
-      const isAdjAbsent = s => s && (s.toLowerCase().startsWith('absent') || s === 'one punch only');
-      const missedAdj = isAdjAbsent(prevStatus) || isAdjAbsent(nextStatus);
-      if (missedAdj) continue; // sandwich unpaid
-
-      const worked = rec && rec.punch_in && rec.punch_out;
-      if (worked) {
-        let hrs = effectiveHours(rec.punch_in, rec.punch_out, 'monthly');
-        if (moment(rec.punch_in, 'HH:mm:ss').isSameOrBefore(graceTime)) {
-          const base = sundayBaseHours;
-          if (hrs < base) hrs = base;
-        }
-        if (emp.pay_sunday) {
-          // Pay double for actual hours worked (after grace adjustment)
-          totalPay += hrs * sundayRate * 2;
-        } else {
-          // Pay base Sunday hours and credit leave when not paid
-          totalPay += sundayBaseHours * sundayRate;
-          const [[row]] = await conn.query(
-            'SELECT id FROM employee_leaves WHERE employee_id = ? AND leave_date = ? LIMIT 1',
-            [employeeId, dateStr]
-          );
-          if (!row) {
-            await conn.query(
-              'INSERT INTO employee_leaves (employee_id, leave_date, days, remark) VALUES (?, ?, 1, ?)',
-              [employeeId, dateStr, 'Sunday Credit']
-            );
-          }
-        }
-      } else {
-        // Absent on Sunday but still paid the base hours
-        totalPay += sundayBaseHours * sundayRate;
-      }
-    } else if (rec && rec.punch_in && rec.punch_out) {
-      let hrs = effectiveHours(rec.punch_in, rec.punch_out, 'monthly');
-      if (moment(rec.punch_in, 'HH:mm:ss').isSameOrBefore(graceTime)) {
-        const base = parseFloat(emp.allotted_hours || 0);
-        if (base && hrs < base) hrs = base;
-      }
-      if (hrs > 0) totalPay += hrs * hourlyRate;
-    }
-  }
-  const gross = parseFloat(totalPay.toFixed(2));
-  const [[advRow]] = await conn.query(
-    'SELECT COALESCE(SUM(amount),0) AS total FROM advance_deductions WHERE employee_id = ? AND month = ?',
-    [employeeId, month]
-  );
-  const advDeduct = parseFloat(advRow.total) || 0;
-  const net = gross - advDeduct;
-  await conn.query(
-    `INSERT INTO employee_salaries (employee_id, month, gross, deduction, net, created_at)
-     VALUES (?, ?, ?, ?, ?, NOW())
-     ON DUPLICATE KEY UPDATE gross=VALUES(gross), deduction=VALUES(deduction), net=VALUES(net)`,
-    [employeeId, month, gross, advDeduct, net]
-  );
-}
-
-exports.calculateSalaryHourly = calculateHourlyMonthly;
