--- routes/salaryRoutes.js
@@ -1,1185 +1,0 @@
-const express = require('express');
-const router = express.Router();
-const multer = require('multer');
-const path = require('path');
-const moment = require('moment');
-const { pool } = require('../config/db');
-const { isAuthenticated, isOperator, isSupervisor } = require('../middlewares/auth');
-const { calculateSalaryForMonth, effectiveHours, lunchDeduction, crossedLunch } = require('../helpers/salaryCalculator');
-const { applyDetailedStatus } = require('../helpers/detailedStatus');
-const { SPECIAL_DEPARTMENTS } = require('../utils/departments');
-const {
-  SPECIAL_SUNDAY_SUPERVISORS,
-  FULL_SALARY_EMPLOYEE_IDS
-} = require('../utils/supervisors');
-const { PRIVILEGED_OPERATOR_ID } = require('../utils/operators');
-
-function formatHours(h) {
-  let hours = Math.floor(h);
-  let mins = Math.round((h - hours) * 60);
-  if (mins === 60) { hours += 1; mins = 0; }
-  return `${String(hours).padStart(2, '0')}:${String(mins).padStart(2, '0')}`;
-}
-
-const { validateAttendanceFilename } = require('../helpers/attendanceFilenameValidator');
-const XLSX = require('xlsx');
-const ExcelJS = require('exceljs');
-
-
-// Configure upload for JSON files in memory
-const upload = multer({ storage: multer.memoryStorage() });
-
-// GET form to upload attendance JSON
-router.get('/salary/upload', isAuthenticated, isOperator, (req, res) => {
-  res.redirect('/operator/departments');
-});
-
-// POST process uploaded attendance JSON
-router.post('/salary/upload', isAuthenticated, isOperator, upload.single('attFile'), async (req, res) => {
-  const file = req.file;
-  if (!file) {
-    req.flash('error', 'No file uploaded');
-    return res.redirect('/operator/departments');
-  }
-
-  const validation = await validateAttendanceFilename(file.originalname);
-  if (!validation.valid) {
-    req.flash('error', validation.message);
-    return res.redirect('/operator/departments');
-  }
-  const supervisorId = validation.supervisorId;
-
-  let data;
-  try {
-    const jsonStr = file.buffer.toString('utf8');
-    data = JSON.parse(jsonStr);
-  } catch (err) {
-    console.error('Failed to parse JSON:', err);
-    req.flash('error', 'Invalid JSON');
-    return res.redirect('/operator/departments');
-  }
-  const conn = await pool.getConnection();
-  try {
-    await conn.beginTransaction();
-    let uploadedCount = 0;
-    const recalc = new Map();
-
-    const [empList] = await conn.query(
-      'SELECT id, punching_id, name FROM employees WHERE supervisor_id = ?',
-      [supervisorId]
-    );
-    const empMap = new Map();
-    empList.forEach(e => {
-      empMap.set(`${e.punching_id}|${(e.name || '').toLowerCase()}`, e.id);
-    });
-
-    for (const emp of data) {
-      const empId = empMap.get(`${emp.punchingId}|${(emp.name || '').toLowerCase()}`);
-      if (!empId) continue;
-
-      const attendanceValues = emp.attendance.map(att => [
-        empId,
-        att.date,
-        att.punchIn || null,
-        att.punchOut || null,
-        att.status || 'present'
-      ]);
-      if (attendanceValues.length) {
-        await conn.query(
-          `INSERT INTO employee_attendance (employee_id, date, punch_in, punch_out, status)
-           VALUES ?
-           ON DUPLICATE KEY UPDATE punch_in = VALUES(punch_in), punch_out = VALUES(punch_out), status = VALUES(status)`,
-          [attendanceValues]
-        );
-      }
-
-      const month = moment(emp.attendance[0].date).format('YYYY-MM');
-      recalc.set(`${empId}_${month}`, { id: empId, month });
-      uploadedCount++;
-    }
-    for (const { id, month } of recalc.values()) {
-      await calculateSalaryForMonth(conn, id, month);
-    }
-    await conn.commit();
-    req.flash('success', `Attendance uploaded for ${uploadedCount} employees`);
-  } catch (err) {
-    await conn.rollback();
-    console.error('Error processing attendance:', err);
-    req.flash('error', 'Failed to process attendance');
-  } finally {
-    conn.release();
-  }
-
-  res.redirect('/operator/departments');
-});
-
-// POST night shift Excel upload
-router.post('/salary/upload-nights', isAuthenticated, isOperator, upload.single('excelFile'), async (req, res) => {
-  const file = req.file;
-  if (!file) {
-    req.flash('error', 'No file uploaded');
-    return res.redirect('/operator/departments');
-  }
-
-  let rows;
-  try {
-    const wb = XLSX.read(file.buffer, { type: 'buffer' });
-    const ws = wb.Sheets[wb.SheetNames[0]];
-    rows = XLSX.utils.sheet_to_json(ws, { defval: '' });
-  } catch (err) {
-    console.error('Failed to parse Excel:', err);
-    req.flash('error', 'Invalid Excel file');
-    return res.redirect('/operator/departments');
-  }
-
-
-  // Night records can be uploaded for any month as long as the employee
-  // already has attendance entries recorded for that month.
-
-  const conn = await pool.getConnection();
-  try {
-    await conn.beginTransaction();
-    let uploadedCount = 0;
-    const recalc = new Map();
-    const supCache = new Map();
-    const empCache = new Map();
-    for (const r of rows) {
-      const month = String(r.month || r.Month || '').trim();
-      if (!month) continue;
-
-      const punchingId = String(r.punchingid || r.punchingId || r.punching_id || '').trim();
-      const name = String(r.name || r.employee_name || r.EmployeeName || '').trim();
-      const supName = String(r.supervisorname || r.supervisor_name || '').trim();
-      const nights = parseInt(r.nights || r.Nights || r.night || 0, 10);
-      if (!punchingId || !name || !nights) continue;
-      let supId = null;
-      if (supName) {
-        if (supCache.has(supName)) {
-          supId = supCache.get(supName);
-        } else {
-          const [[sup]] = await conn.query('SELECT id FROM users WHERE username = ? LIMIT 1', [supName]);
-          supId = sup ? sup.id : null;
-          supCache.set(supName, supId);
-        }
-      }
-      const empKey = `${punchingId}|${name.toLowerCase()}|${supId || ''}`;
-      let empId = empCache.get(empKey);
-      if (empId === undefined) {
-        let supervisorCondition = '';
-        const params = [punchingId, name];
-        if (supId) {
-          supervisorCondition = ' AND supervisor_id = ?';
-          params.push(supId);
-        }
-        const [empRows] = await conn.query(
-          `SELECT id FROM employees WHERE punching_id = ? AND name = ?${supervisorCondition} LIMIT 1`,
-          params
-        );
-        empId = empRows.length ? empRows[0].id : null;
-        empCache.set(empKey, empId);
-      }
-      if (!empId) continue;
-      const [[attMonth]] = await conn.query(
-        'SELECT 1 FROM employee_attendance WHERE employee_id = ? AND DATE_FORMAT(date, "%Y-%m") = ? LIMIT 1',
-        [empId, month]
-      );
-      if (!attMonth) continue;
-
-      const [existing] = await conn.query(
-        'SELECT id FROM employee_nights WHERE employee_id = ? AND month = ? LIMIT 1',
-        [empId, month]
-      );
-      if (existing.length) continue;
-
-      await conn.query(
-        'INSERT INTO employee_nights (employee_id, supervisor_name, supervisor_department, punching_id, employee_name, nights, month) VALUES (?, ?, ?, ?, ?, ?, ?)',
-        [
-          empId,
-          r.supervisorname || r.supervisor_name || '',
-          r.supervisordepartment || r.department || '',
-          punchingId,
-          name,
-          nights,
-          month
-        ]
-      );
-      recalc.set(`${empId}_${month}`, { id: empId, month });
-
-      uploadedCount++;
-    }
-    for (const { id, month } of recalc.values()) {
-      await calculateSalaryForMonth(conn, id, month);
-    }
-    await conn.commit();
-    req.flash('success', `Night data uploaded for ${uploadedCount} employees`);
-  } catch (err) {
-    await conn.rollback();
-    console.error('Error processing night data:', err);
-    req.flash('error', 'Failed to process night data');
-  } finally {
-    conn.release();
-  }
-
-  res.redirect('/operator/departments');
-});
-
-
-// GET night shift Excel template
-router.get('/salary/night-template', isAuthenticated, isOperator, async (req, res) => {
-  try {
-    const workbook = new ExcelJS.Workbook();
-    const sheet = workbook.addWorksheet('NightTemplate');
-    sheet.columns = [
-      { header: 'supervisorname', key: 'supervisorname', width: 20 },
-      { header: 'supervisordepartment', key: 'supervisordepartment', width: 20 },
-      { header: 'punchingid', key: 'punchingid', width: 15 },
-      { header: 'name', key: 'name', width: 20 },
-      { header: 'nights', key: 'nights', width: 10 },
-      { header: 'month', key: 'month', width: 12 }
-    ];
-    res.setHeader('Content-Disposition', 'attachment; filename="NightShiftTemplate.xlsx"');
-    res.setHeader('Content-Type', 'application/vnd.openxmlformats-officedocument.spreadsheetml.sheet');
-    await workbook.xlsx.write(res);
-    return res.end();
-  } catch (err) {
-    console.error('Error downloading night template:', err);
-    req.flash('error', 'Error downloading night template');
-    return res.redirect('/operator/departments');
-  }
-});
-
-// GET advance Excel template
-router.get('/salary/advance-template', isAuthenticated, isOperator, async (req, res) => {
-  try {
-    const workbook = new ExcelJS.Workbook();
-    const sheet = workbook.addWorksheet('AdvanceTemplate');
-    sheet.columns = [
-      { header: 'employeeid', key: 'employeeid', width: 12 },
-      { header: 'punchingid', key: 'punchingid', width: 15 },
-      { header: 'name', key: 'name', width: 20 },
-      { header: 'amount', key: 'amount', width: 10 },
-      { header: 'reason', key: 'reason', width: 20 }
-    ];
-    res.setHeader('Content-Disposition', 'attachment; filename="AdvanceTemplate.xlsx"');
-    res.setHeader('Content-Type', 'application/vnd.openxmlformats-officedocument.spreadsheetml.sheet');
-    await workbook.xlsx.write(res);
-    return res.end();
-  } catch (err) {
-    console.error('Error downloading advance template:', err);
-    req.flash('error', 'Error downloading advance template');
-    return res.redirect('/operator/departments');
-  }
-});
-
-// GET advance deduction Excel template
-router.get('/salary/advance-deduction-template', isAuthenticated, isOperator, async (req, res) => {
-  try {
-    const workbook = new ExcelJS.Workbook();
-    const sheet = workbook.addWorksheet('AdvanceDeductionTemplate');
-    sheet.columns = [
-      { header: 'employeeid', key: 'employeeid', width: 12 },
-      { header: 'punchingid', key: 'punchingid', width: 15 },
-      { header: 'name', key: 'name', width: 20 },
-      { header: 'month', key: 'month', width: 10 },
-      { header: 'amount', key: 'amount', width: 10 }
-    ];
-    res.setHeader('Content-Disposition', 'attachment; filename="AdvanceDeductionTemplate.xlsx"');
-    res.setHeader('Content-Type', 'application/vnd.openxmlformats-officedocument.spreadsheetml.sheet');
-    await workbook.xlsx.write(res);
-    return res.end();
-  } catch (err) {
-    console.error('Error downloading advance deduction template:', err);
-    req.flash('error', 'Error downloading advance deduction template');
-    return res.redirect('/operator/departments');
-  }
-});
-
-// POST advance Excel upload
-router.post('/salary/upload-advances', isAuthenticated, isOperator, upload.single('excelFile'), async (req, res) => {
-  const file = req.file;
-  if (!file) {
-    req.flash('error', 'No file uploaded');
-    return res.redirect('/operator/departments');
-  }
-
-  let rows;
-  try {
-    const wb = XLSX.read(file.buffer, { type: 'buffer' });
-    const ws = wb.Sheets[wb.SheetNames[0]];
-    rows = XLSX.utils.sheet_to_json(ws, { defval: '' });
-  } catch (err) {
-    console.error('Failed to parse Excel:', err);
-    req.flash('error', 'Invalid Excel file');
-    return res.redirect('/operator/departments');
-  }
-
-  const conn = await pool.getConnection();
-  try {
-    await conn.beginTransaction();
-    let uploadedCount = 0;
-    const empCache = new Map();
-    for (const r of rows) {
-      const empIdInput = parseInt(r.employeeid || r.employee_id || r.empid || r.id || 0, 10);
-      let empId = null;
-      if (empIdInput) {
-        if (empCache.has(`id_${empIdInput}`)) {
-          empId = empCache.get(`id_${empIdInput}`);
-        } else {
-          const [[emp]] = await conn.query('SELECT id FROM employees WHERE id = ? LIMIT 1', [empIdInput]);
-          empId = emp ? emp.id : null;
-          empCache.set(`id_${empIdInput}`, empId);
-        }
-      }
-      if (!empId) {
-        const punchingId = String(r.punchingid || r.punchingId || r.punching_id || '').trim();
-        const name = String(r.name || r.employee_name || '').trim();
-        if (!punchingId || !name) continue;
-        const key = `${punchingId}|${name.toLowerCase()}`;
-        if (empCache.has(key)) {
-          empId = empCache.get(key);
-        } else {
-          const [[emp]] = await conn.query('SELECT id FROM employees WHERE punching_id = ? AND name = ? LIMIT 1', [punchingId, name]);
-          empId = emp ? emp.id : null;
-          empCache.set(key, empId);
-        }
-        if (!empId) continue;
-      }
-      const amount = parseFloat(r.amount || r.advance || 0);
-      if (!amount) continue;
-      await conn.query('INSERT INTO employee_advances (employee_id, amount, reason) VALUES (?, ?, ?)', [empId, amount, r.reason || null]);
-      uploadedCount++;
-    }
-    await conn.commit();
-    req.flash('success', `Advance data uploaded for ${uploadedCount} employees`);
-  } catch (err) {
-    await conn.rollback();
-    console.error('Error processing advance data:', err);
-    req.flash('error', 'Failed to process advance data');
-  } finally {
-    conn.release();
-  }
-
-  res.redirect('/operator/departments');
-});
-
-// POST advance deduction Excel upload
-router.post('/salary/upload-advance-deductions', isAuthenticated, isOperator, upload.single('excelFile'), async (req, res) => {
-  const file = req.file;
-  if (!file) {
-    req.flash('error', 'No file uploaded');
-    return res.redirect('/operator/departments');
-  }
-
-  let rows;
-  try {
-    const wb = XLSX.read(file.buffer, { type: 'buffer' });
-    const ws = wb.Sheets[wb.SheetNames[0]];
-    rows = XLSX.utils.sheet_to_json(ws, { defval: '' });
-  } catch (err) {
-    console.error('Failed to parse Excel:', err);
-    req.flash('error', 'Invalid Excel file');
-    return res.redirect('/operator/departments');
-  }
-
-  const conn = await pool.getConnection();
-  try {
-    await conn.beginTransaction();
-    let uploadedCount = 0;
-    const empCache = new Map();
-    for (const r of rows) {
-      const empIdInput = parseInt(r.employeeid || r.employee_id || r.empid || r.id || 0, 10);
-      let empId = null;
-      if (empIdInput) {
-        if (empCache.has(`id_${empIdInput}`)) {
-          empId = empCache.get(`id_${empIdInput}`);
-        } else {
-          const [[emp]] = await conn.query('SELECT id FROM employees WHERE id = ? LIMIT 1', [empIdInput]);
-          empId = emp ? emp.id : null;
-          empCache.set(`id_${empIdInput}`, empId);
-        }
-      }
-      let punchingId = '';
-      let name = '';
-      if (!empId) {
-        punchingId = String(r.punchingid || r.punchingId || r.punching_id || '').trim();
-        name = String(r.name || r.employee_name || '').trim();
-        if (!punchingId || !name) continue;
-        const key = `${punchingId}|${name.toLowerCase()}`;
-        if (empCache.has(key)) {
-          empId = empCache.get(key);
-        } else {
-          const [[emp]] = await conn.query('SELECT id FROM employees WHERE punching_id = ? AND name = ? LIMIT 1', [punchingId, name]);
-          empId = emp ? emp.id : null;
-          empCache.set(key, empId);
-        }
-        if (!empId) continue;
-      }
-      const month = String(r.month || '').trim();
-      const amount = parseFloat(r.amount || r.deduction || 0);
-      if (!month || !amount) continue;
-      const [[sal]] = await conn.query('SELECT id, gross, deduction FROM employee_salaries WHERE employee_id = ? AND month = ? LIMIT 1', [empId, month]);
-      if (!sal) continue;
-      const [[exists]] = await conn.query('SELECT id FROM advance_deductions WHERE employee_id = ? AND month = ? LIMIT 1', [empId, month]);
-      if (exists) continue;
-      const newDed = parseFloat(sal.deduction) + amount;
-      const net = parseFloat(sal.gross) - newDed;
-      await conn.query('UPDATE employee_salaries SET deduction = ?, net = ? WHERE id = ?', [newDed, net, sal.id]);
-      await conn.query('INSERT INTO advance_deductions (employee_id, month, amount) VALUES (?, ?, ?)', [empId, month, amount]);
-      uploadedCount++;
-    }
-    await conn.commit();
-    req.flash('success', `Advance deductions uploaded for ${uploadedCount} employees`);
-  } catch (err) {
-    await conn.rollback();
-    console.error('Error processing advance deductions:', err);
-    req.flash('error', 'Failed to process advance deductions');
-  } finally {
-    conn.release();
-  }
-
-  res.redirect('/operator/departments');
-});
-
-
-// View salary summary for operator
-router.get('/salaries', isAuthenticated, isOperator, async (req, res) => {
-  try {
-    const [rows] = await pool.query(`
-      SELECT u.username AS supervisor_name, u.id AS supervisor_id,
-             COUNT(e.id) AS employee_count,
-             SUM(CASE WHEN e.is_active = 1 THEN e.salary ELSE 0 END) AS total_salary
-        FROM users u
-        JOIN employees e ON e.supervisor_id = u.id
-       GROUP BY u.id`);
-    const canViewSalary = req.session.user.id === PRIVILEGED_OPERATOR_ID;
-    res.render('operatorSalaries', {
-      user: req.session.user,
-      summary: rows,
-      canViewSalary
-    });
-  } catch (err) {
-    console.error('Error loading salary summary:', err);
-    req.flash('error', 'Could not load salary summary');
-    res.redirect('/dashboard');
-  }
-
-});
-
-// Supervisor view of employee salary
-router.get('/employees/:id/salary', isAuthenticated, isSupervisor, async (req, res) => {
-  const empId = req.params.id;
-  const month = req.query.month || moment().format('YYYY-MM');
-  const half = parseInt(req.query.half, 10) === 2 ? 2 : 1;
-  const hourlyView = req.query.hourly === '1';
-  try {
-    const [[emp]] = await pool.query(
-      `SELECT e.*, d.name AS department
-         FROM employees e
-         LEFT JOIN (
-               SELECT user_id, MIN(department_id) AS department_id
-                 FROM department_supervisors
-                GROUP BY user_id
-         ) ds ON ds.user_id = e.supervisor_id
-         LEFT JOIN departments d ON ds.department_id = d.id
-        WHERE e.id = ? AND e.supervisor_id = ?`,
-      [empId, req.session.user.id]
-    );
-    if (!emp) {
-      req.flash('error', 'Employee not found');
-      return res.redirect('/supervisor/employees');
-    }
-    const manualSalary = (emp.designation || '').toLowerCase() === 'checker';
-    const specialDept = SPECIAL_DEPARTMENTS.includes(
-      (emp.department || '').toLowerCase()
-    );
-    let startDate = moment(month + '-01');
-    let endDate;
-    if (emp.salary_type === 'dihadi') {
-      if (half === 2) {
-        startDate = moment(month + '-16');
-        endDate = moment(month + '-01').endOf('month');
-      } else {
-        endDate = moment(month + '-15');
-      }
-    } else {
-      endDate = moment(month + '-01').endOf('month');
-    }
-
-    if (emp.date_of_joining) {
-      const joinDate = moment(emp.date_of_joining);
-      if (joinDate.isAfter(startDate)) {
-        startDate = joinDate;
-      }
-    }
-
-    const startStr = startDate.format('YYYY-MM-DD');
-    const endStr = endDate.format('YYYY-MM-DD');
-    const [attendance] = await pool.query(
-      'SELECT * FROM employee_attendance WHERE employee_id = ? AND date BETWEEN ? AND ? ORDER BY date',
-      [empId, startStr, endStr]
-    );
-    const [sandwichRows] = await pool.query(
-      'SELECT date FROM sandwich_dates WHERE DATE_FORMAT(date, "%Y-%m") = ?',
-      [month]
-    );
-    const sandwichDates = sandwichRows.map(r =>
-      moment(r.date).format('YYYY-MM-DD')
-    );
-    applyDetailedStatus(attendance, emp, sandwichDates);
-    const daysInMonth = moment(month + '-01').daysInMonth();
-    let dailyRate;
-    if (emp.salary_type === 'dihadi') {
-      dailyRate = parseFloat(emp.salary);
-    } else {
-      dailyRate = parseFloat(emp.salary) / daysInMonth;
-    }
-    let totalHours = 0;
-    let sundayHours = 0;
-    let hourlyRate = 0;
-    const sundayBaseHours = 9;
-    const sundayRate = dailyRate / sundayBaseHours;
-    let partialPay = 0;
-    let overtimeTotal = 0;
-    let undertimeTotal = 0;
-    if (emp.salary_type === 'dihadi') {
-      hourlyRate = emp.allotted_hours
-        ? parseFloat(emp.salary) / parseFloat(emp.allotted_hours)
-        : 0;
-    } else if (emp.salary_type === 'monthly' && hourlyView) {
-      hourlyRate = emp.allotted_hours
-        ? dailyRate / parseFloat(emp.allotted_hours)
-        : 0;
-    }
-    attendance.forEach(a => {
-      const isSun = moment(a.date).day() === 0;
-      let hrsDec = 0;
-      if (a.punch_in && a.punch_out) {
-        hrsDec = effectiveHours(
-          a.punch_in,
-          a.punch_out,
-          emp.salary_type,
-          emp.salary_type === 'dihadi' ? emp.allotted_hours : undefined
-        );
-        a.hours = formatHours(hrsDec);
-        a.lunch_deduction = lunchDeduction(
-          a.punch_in,
-          a.punch_out,
-          emp.salary_type,
-          emp.salary_type === 'dihadi' ? emp.allotted_hours : undefined
-        );
-        if (emp.salary_type === 'monthly') {
-          const baseHours = isSun ? 9 : parseFloat(emp.allotted_hours || 0);
-          let diff = hrsDec - baseHours;
-          if (
-            hourlyView &&
-            diff < 0 &&
-            moment(a.punch_in, 'HH:mm:ss').isSameOrBefore(moment('09:15:00', 'HH:mm:ss'))
-          ) {
-            diff = 0; // grace minutes when punching in before 09:15
-          }
-          if (diff > 0) {
-            a.overtime = formatHours(diff);
-            a.undertime = '00:00';
-            overtimeTotal += diff;
-          } else if (diff < 0 && crossedLunch(a.punch_in, a.punch_out)) {
-            a.overtime = '00:00';
-            a.undertime = formatHours(Math.abs(diff));
-            undertimeTotal += Math.abs(diff);
-          } else if (diff < 0) {
-            a.overtime = '00:00';
-            a.undertime = '00:00';
-          } else {
-            a.overtime = '00:00';
-            a.undertime = '00:00';
-          }
-          totalHours += hrsDec;
-          if (isSun) sundayHours += hrsDec;
-        } else {
-          totalHours += hrsDec;
-          if (moment(a.date).day() === 0) sundayHours += hrsDec;
-        }
-        // amount will be calculated after determining detailed status
-      } else {
-        a.hours = '00:00';
-        a.lunch_deduction = 0;
-        if (emp.salary_type === 'monthly') {
-          a.overtime = '00:00';
-          a.undertime = '00:00';
-        }
-        // amount will be calculated after determining detailed status
-      }
-      const status = a.detailed_status || a.status;
-      let reason = '';
-      if (/^Absent/.test(status)) {
-        if (status.includes('Sandwich') || status.includes('Mandatory')) {
-          reason = 'Sunday absence';
-        } else if (status.includes('Short hours')) {
-          reason = 'Short hours absence';
-        } else {
-          reason = 'Absent from work';
-        }
-      } else if (status === 'Missing punch') {
-        reason = 'Missing punch in/out';
-      } else if (status === 'Half Day') {
-        reason = 'Worked less than half day';
-      } else if (status === 'Paid Sunday' || status === 'Paid due to Sunday work') {
-        reason = 'Paid Sunday';
-      } else if (status === 'Leave credited') {
-        reason = 'Leave credited';
-      }
-      if (
-        emp.salary_type === 'dihadi' &&
-        a.punch_in &&
-        moment(a.punch_in, 'HH:mm:ss').isAfter(moment('09:15:00', 'HH:mm:ss'))
-      ) {
-        reason += (reason ? '; ' : '') + 'Late arrival after 09:15';
-      }
-      a.deduction_reason = reason;
-      if (emp.salary_type === 'monthly' && !hourlyView) {
-        if (/^Absent/.test(status) || status === 'Missing punch') {
-          a.amount = isSun ? parseFloat(dailyRate.toFixed(2)) : 0;
-        } else if (status === 'Half Day') {
-          a.amount = parseFloat((dailyRate / 2).toFixed(2));
-        } else if (status === 'Paid Sunday') {
-          const hoursForPay = hrsDec || sundayBaseHours;
-          a.amount = parseFloat((hoursForPay * sundayRate * 2).toFixed(2));
-        } else if (status === 'Paid due to Sunday work') {
-          a.amount = parseFloat(dailyRate.toFixed(2));
-        } else {
-          a.amount = parseFloat(dailyRate.toFixed(2));
-        }
-      } else if (emp.salary_type === 'monthly' && hourlyView) {
-        let hrsPay = hrsDec;
-        if (
-          moment(a.punch_in, 'HH:mm:ss').isSameOrBefore(moment('09:15:00', 'HH:mm:ss')) &&
-          hrsDec < (isSun ? sundayBaseHours : parseFloat(emp.allotted_hours || 0))
-        ) {
-          hrsPay = isSun ? sundayBaseHours : parseFloat(emp.allotted_hours || 0);
-        }
-        let amt;
-        if (status === 'Leave credited') {
-          amt = sundayBaseHours * sundayRate;
-        } else if (status.startsWith('Absent (Sandwich)') || status.startsWith('Absent (Mandatory)')) {
-          amt = 0;
-        } else if (isSun) {
-          amt = hrsPay * sundayRate * 2;
-        } else {
-          amt = hrsPay * hourlyRate;
-        }
-        a.amount = parseFloat(amt.toFixed(2));
-        partialPay += amt;
-      } else if (emp.salary_type === 'dihadi') {
-        const amt = hrsDec * hourlyRate;
-        a.amount = parseFloat(amt.toFixed(2));
-        partialPay += amt;
-      }
-    });
-    let totalHoursFormatted = null;
-    let sundayHoursFormatted = null;
-    if (emp.salary_type === 'dihadi' || emp.salary_type === 'monthly') {
-      totalHoursFormatted = formatHours(totalHours);
-      sundayHoursFormatted = formatHours(sundayHours);
-    }
-    let overtimeFormatted = null;
-    let undertimeFormatted = null;
-    if (emp.salary_type === 'monthly') {
-      overtimeFormatted = formatHours(overtimeTotal);
-      undertimeFormatted = formatHours(undertimeTotal);
-    }
-    let partialAmount = null;
-    if (emp.salary_type === 'dihadi') {
-      partialAmount = parseFloat((totalHours * hourlyRate).toFixed(2));
-    } else if (emp.salary_type === 'monthly' && hourlyView) {
-      partialAmount = parseFloat(partialPay.toFixed(2));
-    }
-    const [[[salary]], [[adv]], [[ded]]] = await Promise.all([
-      pool.query('SELECT * FROM employee_salaries WHERE employee_id = ? AND month = ? LIMIT 1', [empId, month]),
-      pool.query('SELECT COALESCE(SUM(amount),0) AS total FROM employee_advances WHERE employee_id = ?', [empId]),
-      pool.query('SELECT COALESCE(SUM(amount),0) AS total FROM advance_deductions WHERE employee_id = ?', [empId])
-    ]);
-    const outstanding = parseFloat(adv.total) - parseFloat(ded.total);
-    res.render('employeeSalary', {
-      user: req.session.user,
-      employee: emp,
-      attendance,
-      salary,
-      manualSalary,
-      month,
-      dailyRate,
-      totalHours: totalHoursFormatted,
-      sundayHours: sundayHoursFormatted,
-      hourlyRate,
-      half,
-      outstanding,
-      overtimeFormatted,
-      undertimeFormatted,
-      partialAmount,
-      hourlyMode: hourlyView
-    });
-  } catch (err) {
-    console.error('Error loading salary view:', err);
-    req.flash('error', 'Failed to load salary');
-  res.redirect('/supervisor/employees');
-  }
-});
-
-// Supervisor download of employee salary sheet
-router.get('/supervisor/salary/download', isAuthenticated, isSupervisor, async (req, res) => {
-  const month = req.query.month || moment().format('YYYY-MM');
-  const supervisorId = req.session.user.id;
-  const daysInMonth = moment(month + '-01').daysInMonth();
-  let totalSundays = 0;
-  for (let d = 1; d <= daysInMonth; d++) {
-    if (moment(`${month}-${String(d).padStart(2, '0')}`).day() === 0) totalSundays++;
-  }
-  try {
-    const [sandwichRows] = await pool.query(
-      'SELECT date FROM sandwich_dates WHERE DATE_FORMAT(date, "%Y-%m") = ?',
-      [month]
-    );
-    const sandwichDates = sandwichRows.map(r => moment(r.date).format('YYYY-MM-DD'));
-
-    const [rows] = await pool.query(`
-      SELECT es.employee_id, es.gross, es.deduction, es.net, es.month,
-             e.punching_id, e.name AS employee_name, e.aadhar_card_number, e.salary AS base_salary, e.salary_type,
-             e.paid_sunday_allowance, e.pay_sunday, e.allotted_hours,
-             (SELECT COALESCE(SUM(amount),0) FROM employee_advances ea WHERE ea.employee_id = es.employee_id) AS advance_taken,
-             (SELECT COALESCE(SUM(amount),0) FROM advance_deductions ad WHERE ad.employee_id = es.employee_id) AS advance_deducted,
-             (SELECT COALESCE(SUM(amount),0) FROM advance_deductions ad WHERE ad.employee_id = es.employee_id AND ad.month = es.month) AS month_ded,
-             (SELECT COALESCE(SUM(nights),0) FROM employee_nights en WHERE en.employee_id = es.employee_id AND en.month = es.month) AS nights,
-             u.username AS supervisor_name, d.name AS department_name
-        FROM employee_salaries es
-        JOIN employees e ON es.employee_id = e.id
-        JOIN users u ON e.supervisor_id = u.id
-        LEFT JOIN (
-              SELECT user_id, MIN(department_id) AS department_id
-                FROM department_supervisors
-               GROUP BY user_id
-        ) ds ON ds.user_id = u.id
-       LEFT JOIN departments d ON ds.department_id = d.id
-       WHERE es.month = ? AND e.is_active = 1 AND e.salary_type = 'monthly' AND e.supervisor_id = ?
-       ORDER BY e.name
-    `, [month, supervisorId]);
-    const empIds = rows.map(r => r.employee_id);
-    const [allAtt] = await pool.query(
-      'SELECT employee_id, date, status, punch_in, punch_out FROM employee_attendance WHERE employee_id IN (?) AND DATE_FORMAT(date, "%Y-%m") = ? ORDER BY employee_id, date',
-      [empIds, month]
-    );
-    const prevDay = moment(month + '-01').subtract(1, 'day').format('YYYY-MM-DD');
-    const nextDay = moment(month + '-01').endOf('month').add(1, 'day').format('YYYY-MM-DD');
-    const [adjacentRows] = await pool.query(
-      'SELECT employee_id, date, status FROM employee_attendance WHERE employee_id IN (?) AND date IN (?, ?)',
-      [empIds, prevDay, nextDay]
-    );
-    const attByEmp = {};
-    allAtt.forEach(a => {
-      const arr = attByEmp[a.employee_id] || (attByEmp[a.employee_id] = []);
-      arr.push(a);
-    });
-    adjacentRows.forEach(a => {
-      const arr = attByEmp[a.employee_id] || (attByEmp[a.employee_id] = []);
-      arr.push(a);
-    });
-
-    for (const r of rows) {
-      const attRows = attByEmp[r.employee_id] || [];
-      const attMap = {};
-      attRows.forEach(a => {
-        attMap[moment(a.date).format('YYYY-MM-DD')] = a.status;
-      });
-      const specialSup =
-        SPECIAL_SUNDAY_SUPERVISORS.map(s => s.toLowerCase()).includes(
-          (r.supervisor_name || '').toLowerCase()
-        ) && !FULL_SALARY_EMPLOYEE_IDS.includes(r.employee_id);
-      let absent = 0, onePunch = 0, sundayAbs = 0;
-      let otHours = 0, utHours = 0, otDays = 0, utDays = 0;
-      let workingDays = 0;
-      let sundaysWorked = 0;
-      const missPunchDates = [];
-      const absentDates = [];
-      attRows.forEach(a => {
-        const dateStr = moment(a.date).format('YYYY-MM-DD');
-        const status = a.status;
-        const isSun = moment(a.date).day() === 0;
-        const isSandwich = !specialSup && sandwichDates.includes(dateStr);
-        let recordedAbsent = false;
-        if (status === 'present' && a.punch_in && a.punch_out) {
-          const hrs = effectiveHours(a.punch_in, a.punch_out, 'monthly');
-          const allot = parseFloat(r.allotted_hours || 0);
-          const half = allot && hrs >= allot * 0.4 && hrs < allot * 0.85;
-          workingDays += half ? 0.5 : 1;
-          if (isSun) sundaysWorked++;
-        } else if (status === 'one punch only') {
-          if (!(specialSup && isSun)) missPunchDates.push(dateStr);
-        } else if (status === 'absent') {
-          if (!(specialSup && isSun)) {
-            absentDates.push(dateStr);
-            recordedAbsent = true;
-          }
-        }
-        if (!specialSup && isSun) {
-          const satKey = moment(a.date).subtract(1, 'day').format('YYYY-MM-DD');
-          const monKey = moment(a.date).add(1, 'day').format('YYYY-MM-DD');
-          const satStatus = attMap[satKey] !== undefined ? attMap[satKey] : 'present';
-          const monStatus = attMap[monKey] !== undefined ? attMap[monKey] : 'present';
-          const adjAbsent =
-            (satStatus === 'absent' || satStatus === 'one punch only') &&
-            (monStatus === 'absent' || monStatus === 'one punch only');
-          if (adjAbsent) {
-            sundayAbs++;
-            if (!recordedAbsent) absentDates.push(dateStr);
-            return;
-          }
-        } else if (specialSup && isSun) {
-          const satKey = moment(a.date).subtract(1, 'day').format('YYYY-MM-DD');
-          const monKey = moment(a.date).add(1, 'day').format('YYYY-MM-DD');
-          const satStatus = attMap[satKey] !== undefined ? attMap[satKey] : 'present';
-          const monStatus = attMap[monKey] !== undefined ? attMap[monKey] : 'present';
-          const adjAbsent =
-            (satStatus === 'absent' || satStatus === 'one punch only') &&
-            (monStatus === 'absent' || monStatus === 'one punch only');
-          if (adjAbsent) {
-            sundayAbs++;
-            if (!recordedAbsent) absentDates.push(dateStr);
-            absent++; // deduct the Sunday as well
-            return;
-          }
-        }
-        if (isSandwich) {
-          const prevStatus = attMap[moment(a.date).subtract(1, 'day').format('YYYY-MM-DD')];
-          const nextStatus = attMap[moment(a.date).add(1, 'day').format('YYYY-MM-DD')];
-          const adjAbsent = (prevStatus === 'absent' || prevStatus === 'one punch only') ||
-                            (nextStatus === 'absent' || nextStatus === 'one punch only');
-          if (adjAbsent) {
-            absent++;
-            if (!recordedAbsent) absentDates.push(dateStr);
-            return;
-          }
-        }
-        if (!isSun) {
-          if (status === 'absent') absent++;
-          else if (status === 'one punch only') onePunch++;
-        }
-        if (a.punch_in && a.punch_out) {
-          const hrs = effectiveHours(a.punch_in, a.punch_out, 'monthly');
-          const isSun = moment(a.date).day() === 0;
-          const baseHours = isSun ? 9 : parseFloat(r.allotted_hours || 0);
-          const diff = hrs - baseHours;
-          if (diff > 0) { otHours += diff; otDays++; }
-          else if (diff < 0 && crossedLunch(a.punch_in, a.punch_out)) {
-            utHours += Math.abs(diff); utDays++; }
-        }
-      });
-      const notes = [];
-      if (absent) notes.push(`${absent} day(s) absent`);
-      if (onePunch) notes.push(`${onePunch} day(s) with missing punch`);
-      if (sundayAbs) notes.push(`${sundayAbs} Sunday absence(s)`);
-      r.deduction_reason = notes.length ? notes.join(', ') : 'None';
-      r.overtime_hours = otHours.toFixed(2);
-      r.overtime_days = otDays;
-      r.undertime_hours = utHours.toFixed(2);
-      r.undertime_days = utDays;
-      r.time_status = otHours > utHours ? 'Overtime' : utHours > otHours ? 'Undertime' : 'Even';
-
-      const netHours = utHours - otHours;
-      const allot = parseFloat(r.allotted_hours || 0);
-      let utDeduct = 0;
-      let utDetail = '';
-      if (netHours > 0 && allot > 0) {
-        const cutDays = Math.floor(netHours / allot);
-        if (cutDays > 0) {
-          const dailyRate = parseFloat(r.base_salary) / daysInMonth;
-          utDeduct = parseFloat((dailyRate * cutDays).toFixed(2));
-          r.deduction = parseFloat(r.deduction) + utDeduct;
-          r.net = parseFloat(r.net) - utDeduct;
-          utDetail = `${cutDays} day salary cut due to undertime`;
-        }
-      }
-      r.ut_deduct = utDeduct.toFixed(2);
-      r.ut_detail = utDetail;
-      r.working_days = workingDays;
-      r.sunday_worked = sundaysWorked;
-      r.absents = absent;
-      if (specialSup) {
-        r.week_off = Math.max(0, totalSundays - sundaysWorked);
-      } else {
-        r.week_off = Math.max(0, totalSundays - (r.paid_sunday_allowance || 0));
-      }
-      r.miss_punch_dates = missPunchDates; // retained for potential debugging
-    }
-
-    // daysInMonth calculated earlier
-    const workbook = new ExcelJS.Workbook();
-    const sheet = workbook.addWorksheet('Salary');
-    const columns = [
-      { header: 'Punching ID', key: 'punching_id', width: 12 },
-      { header: 'Aadhar', key: 'aadhar', width: 18 },
-      { header: 'Employee', key: 'employee', width: 20 },
-      { header: 'Salary Type', key: 'salary_type', width: 12 },
-      { header: 'Base Salary', key: 'base_salary', width: 12 }
-    ];
-    for (let d = 1; d <= daysInMonth; d++) {
-      const key = `d${String(d).padStart(2, '0')}`;
-      columns.push({ header: String(d).padStart(2, '0'), key, width: 4 });
-    }
-    columns.push({ header: 'Advance Deduct', key: 'advance_deduct', width: 14 });
-    columns.push({ header: 'Net', key: 'net', width: 10 });
-    columns.push({ header: 'UT Deduct', key: 'ut_deduct', width: 12 });
-    columns.push({ header: 'UT Detail', key: 'ut_detail', width: 25 });
-    columns.push({ header: 'Working Days', key: 'working_days', width: 14 });
-    columns.push({ header: 'Sundays Worked', key: 'sunday_worked', width: 15 });
-    columns.push({ header: 'Absents', key: 'absents', width: 10 });
-    columns.push({ header: 'Week Off', key: 'week_off', width: 10 });
-    columns.push({ header: 'Nights', key: 'nights', width: 10 });
-    sheet.columns = columns;
-    sheet.getRow(1).font = { bold: true };
-    sheet.getRow(1).alignment = { horizontal: 'center' };
-    sheet.getRow(1).eachCell(cell => {
-      cell.border = {
-        top: { style: 'thin' },
-        left: { style: 'thin' },
-        bottom: { style: 'thin' },
-        right: { style: 'thin' }
-      };
-    });
-    sheet.views = [{ state: 'frozen', ySplit: 1 }];
-    sheet.getColumn('base_salary').numFmt = '0.00';
-    sheet.getColumn('net').numFmt = '0.00';
-    sheet.getColumn('ut_deduct').numFmt = '0.00';
-    sheet.getColumn('base_salary').alignment = { horizontal: 'right' };
-    sheet.getColumn('net').alignment = { horizontal: 'right' };
-    sheet.getColumn('ut_deduct').alignment = { horizontal: 'right' };
-    sheet.getColumn('working_days').alignment = { horizontal: 'center' };
-    sheet.getColumn('sunday_worked').alignment = { horizontal: 'center' };
-    sheet.getColumn('absents').alignment = { horizontal: 'center' };
-    sheet.getColumn('week_off').alignment = { horizontal: 'center' };
-    sheet.getColumn('nights').alignment = { horizontal: 'center' };
-
-    for (const r of rows) {
-      const attRows = attByEmp[r.employee_id] || [];
-      const attMap = {};
-      attRows.forEach(a => {
-        attMap[moment(a.date).format('YYYY-MM-DD')] = a;
-      });
-      const specialSup =
-        SPECIAL_SUNDAY_SUPERVISORS.map(s => s.toLowerCase()).includes(
-          (r.supervisor_name || '').toLowerCase()
-        ) && !FULL_SALARY_EMPLOYEE_IDS.includes(r.employee_id);
-      const rowData = {
-        punching_id: r.punching_id,
-        aadhar: r.aadhar_card_number || '',
-        employee: r.employee_name,
-        salary_type: r.salary_type,
-        base_salary: r.base_salary
-      };
-      let sundayCounter = 0;
-      for (let d = 1; d <= daysInMonth; d++) {
-        const date = moment(month + '-' + String(d).padStart(2, '0'));
-        const key = `d${String(d).padStart(2, '0')}`;
-        const rec = attMap[date.format('YYYY-MM-DD')];
-        let char = 'A';
-        if (date.day() === 0) sundayCounter++;
-        const mandatory = date.day() === 0 && sundayCounter <= (r.paid_sunday_allowance || 0);
-
-        if (rec && rec.punch_in && rec.punch_out && rec.status === 'present') {
-          const hrs = effectiveHours(rec.punch_in, rec.punch_out, 'monthly');
-          const allot = parseFloat(r.allotted_hours || 0);
-          if (hrs >= allot * 0.4 && hrs < allot * 0.85) {
-            char = 'H';
-          } else {
-            char = hrs.toFixed(2);
-          }
-          if (date.day() === 0 && !mandatory) {
-            if (r.pay_sunday && !specialSup && !SPECIAL_DEPARTMENTS.includes((r.department_name || '').toLowerCase())) {
-              char = 'ED';
-            } else {
-              char = 'WO';
-            }
-          }
-        } else if (date.day() === 0) {
-          char = mandatory ? 'A' : 'WO';
-        }
-        rowData[key] = char;
-      }
-      rowData.advance_deduct = r.month_ded;
-      rowData.net = r.net;
-      rowData.ut_deduct = r.ut_deduct;
-      rowData.ut_detail = r.ut_detail;
-      rowData.working_days = r.working_days;
-      rowData.sunday_worked = r.sunday_worked;
-      rowData.absents = r.absents;
-      rowData.week_off = r.week_off;
-      rowData.nights = r.nights;
-      sheet.addRow(rowData);
-    }
-    res.setHeader('Content-Disposition', 'attachment; filename="SalarySummary.xlsx"');
-    res.setHeader('Content-Type', 'application/vnd.openxmlformats-officedocument.spreadsheetml.sheet');
-    await workbook.xlsx.write(res);
-    res.end();
-  } catch (err) {
-    console.error('Error downloading salary:', err);
-    req.flash('error', 'Could not download salary');
-    res.redirect('/supervisor/employees');
-  }
-});
-
-// Supervisor download of dihadi hours sheet
-router.get('/supervisor/dihadi/download', isAuthenticated, isSupervisor, async (req, res) => {
-  const month = req.query.month || moment().format('YYYY-MM');
-  const half = parseInt(req.query.half, 10) === 2 ? 2 : 1;
-  const supervisorId = req.session.user.id;
-  let start = moment(month + '-01');
-  let end = half === 1 ? moment(month + '-15') : moment(month + '-01').endOf('month');
-  if (half === 2) start = moment(month + '-16');
-  try {
-    const [employees] = await pool.query(
-      `SELECT e.id, e.punching_id, e.name, e.aadhar_card_number, e.salary, e.allotted_hours,
-              (SELECT COALESCE(SUM(amount),0) FROM employee_advances ea WHERE ea.employee_id = e.id) AS advance_taken,
-              (SELECT COALESCE(SUM(amount),0) FROM advance_deductions ad WHERE ad.employee_id = e.id) AS advance_deducted
-         FROM employees e
-        WHERE e.supervisor_id = ? AND e.salary_type = 'dihadi' AND e.is_active = 1
-     ORDER BY e.name`,
-      [supervisorId]
-    );
-    const empIds = employees.map(e => e.id);
-    const rows = [];
-    if (empIds.length) {
-      const [attAll] = await pool.query(
-        'SELECT employee_id, date, punch_in, punch_out FROM employee_attendance WHERE employee_id IN (?) AND date BETWEEN ? AND ? ORDER BY employee_id, date',
-        [empIds, start.format('YYYY-MM-DD'), end.format('YYYY-MM-DD')]
-      );
-      const attMap = new Map();
-      for (const a of attAll) {
-        const key = a.employee_id;
-        if (!attMap.has(key)) attMap.set(key, []);
-        attMap.get(key).push(a);
-      }
-      for (const emp of employees) {
-        const att = attMap.get(emp.id) || [];
-        const byDate = {};
-        let totalHrs = 0;
-        let totalLunch = 0;
-        for (const a of att) {
-          const day = moment(a.date).date();
-          if (a.punch_in && a.punch_out) {
-            const hrs = effectiveHours(
-              a.punch_in,
-              a.punch_out,
-              'dihadi',
-              emp.allotted_hours
-            );
-            const lunch = lunchDeduction(
-              a.punch_in,
-              a.punch_out,
-              'dihadi',
-              emp.allotted_hours
-            );
-            byDate[day] = hrs.toFixed(2);
-            totalHrs += hrs;
-            totalLunch += lunch;
-          } else {
-            byDate[day] = '';
-          }
-        }
-        const row = {
-          employee: emp.name,
-          punching_id: emp.punching_id,
-          aadhar: emp.aadhar_card_number || '',
-          base_amount: emp.salary
-        };
-        for (let d = start.date(); d <= end.date(); d++) {
-          const key = `d${String(d).padStart(2, '0')}`;
-          row[key] = byDate[d] !== undefined ? byDate[d] : '';
-        }
-        const rate = emp.allotted_hours
-          ? parseFloat(emp.salary) / parseFloat(emp.allotted_hours)
-          : 0;
-        const amount = parseFloat((totalHrs * rate).toFixed(2));
-        const net = parseFloat((amount - parseFloat(emp.advance_deducted)).toFixed(2));
-        row.total_hours = totalHrs.toFixed(2);
-        row.lunch_deduction = (totalLunch / 60).toFixed(2);
-        row.total_amount = amount;
-        row.advance_deduct = parseFloat(emp.advance_deducted);
-        row.net_payment = net;
-        rows.push(row);
-      }
-    }
-    const workbook = new ExcelJS.Workbook();
-    const sheet = workbook.addWorksheet('Dihadi');
-    const columns = [
-      { header: 'Employee', key: 'employee', width: 20 },
-      { header: 'Punching ID', key: 'punching_id', width: 12 },
-      { header: 'Aadhar', key: 'aadhar', width: 18 },
-      { header: 'Dihadi Base', key: 'base_amount', width: 12 }
-    ];
-    for (let d = start.date(); d <= end.date(); d++) {
-      const key = `d${String(d).padStart(2, '0')}`;
-      columns.push({ header: String(d), key, width: 5 });
-    }
-    columns.push({ header: 'Total Hours', key: 'total_hours', width: 12 });
-    columns.push({ header: 'Lunch Deduction', key: 'lunch_deduction', width: 15 });
-    columns.push({ header: 'Total Amount', key: 'total_amount', width: 12 });
-    columns.push({ header: 'Advance Deducted', key: 'advance_deduct', width: 15 });
-    columns.push({ header: 'Net Payment', key: 'net_payment', width: 12 });
-    sheet.columns = columns;
-    rows.forEach(r => sheet.addRow(r));
-    res.setHeader('Content-Disposition', 'attachment; filename="DihadiHours.xlsx"');
-    res.setHeader('Content-Type', 'application/vnd.openxmlformats-officedocument.spreadsheetml.sheet');
-    await workbook.xlsx.write(res);
-    res.end();
-  } catch (err) {
-    console.error('Error downloading dihadi hours:', err);
-    req.flash('error', 'Could not download dihadi hours');
-    res.redirect('/supervisor/employees');
-  }
-});
-
-router.post('/employees/:id/salary/deduct-advance', isAuthenticated, isSupervisor, async (req, res) => {
-  const empId = req.params.id;
-  const { month, amount } = req.body;
-  const conn = await pool.getConnection();
-  try {
-    await conn.beginTransaction();
-    const [[latest]] = await conn.query('SELECT month, id, gross, deduction FROM employee_salaries WHERE employee_id = ? ORDER BY month DESC LIMIT 1', [empId]);
-    if (!latest || latest.month !== month) {
-      req.flash('error', 'Can only deduct from the latest salary record');
-      await conn.rollback();
-      conn.release();
-      return res.redirect(`/employees/${empId}/salary?month=${month}`);
-    }
-    const [[exists]] = await conn.query('SELECT id FROM advance_deductions WHERE employee_id = ? AND month = ? LIMIT 1', [empId, month]);
-    if (exists) {
-      req.flash('error', 'Advance already deducted for this salary');
-      await conn.rollback();
-      conn.release();
-      return res.redirect(`/employees/${empId}/salary?month=${month}`);
-    }
-    const [[[adv]], [[ded]]] = await Promise.all([
-      conn.query('SELECT COALESCE(SUM(amount),0) AS total FROM employee_advances WHERE employee_id = ?', [empId]),
-      conn.query('SELECT COALESCE(SUM(amount),0) AS total FROM advance_deductions WHERE employee_id = ?', [empId])
-    ]);
-    const outstanding = parseFloat(adv.total) - parseFloat(ded.total);
-    const amt = parseFloat(amount);
-    if (amt <= 0 || amt > outstanding) {
-      req.flash('error', 'Invalid deduction amount');
-      await conn.rollback();
-      conn.release();
-      return res.redirect(`/employees/${empId}/salary?month=${month}`);
-    }
-    const newDed = parseFloat(latest.deduction) + amt;
-    const net = parseFloat(latest.gross) - newDed;
-    await conn.query('UPDATE employee_salaries SET deduction = ?, net = ? WHERE id = ?', [newDed, net, latest.id]);
-    await conn.query('INSERT INTO advance_deductions (employee_id, month, amount) VALUES (?, ?, ?)', [empId, month, amt]);
-    await conn.commit();
-    req.flash('success', 'Advance deducted');
-  } catch (err) {
-    await conn.rollback();
-    console.error('Error deducting advance:', err);
-    req.flash('error', 'Failed to deduct advance');
-  } finally {
-    conn.release();
-  }
-  res.redirect(`/employees/${empId}/salary?month=${month}`);
-});
-
-module.exports = router;
