--- views/employeeSalary.ejs
@@ -1,141 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-  <meta charset="UTF-8">
-  <title>Employee Salary</title>
-  <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.3.2/dist/css/bootstrap.min.css" rel="stylesheet">
-  <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/6.5.1/css/all.min.css">
-  <link href="https://fonts.googleapis.com/css2?family=Roboto&display=swap" rel="stylesheet">
-  <style>
-    body { font-family: 'Roboto', sans-serif; font-size: 0.95rem; }
-    .toggle-salary { cursor: pointer; }
-  </style>
-</head>
-<body>
-<nav class="navbar navbar-dark bg-dark">
-  <div class="container-fluid">
-    <span class="navbar-brand"><%= employee.name %> (<%= employee.designation || 'N/A' %>) - Salary</span>
-    <div class="ms-auto">
-      <a href="/supervisor/employees" class="btn btn-outline-light btn-sm me-2">Back</a>
-      <a href="/logout" class="btn btn-outline-light btn-sm">Logout</a>
-    </div>
-  </div>
-</nav>
-<div class="container-fluid my-4">
-  <%- include('partials/flashMessages') %>
-  <form method="GET" action="/employees/<%= employee.id %>/salary" class="row g-2 mb-3">
-    <div class="col-auto">
-      <input type="month" name="month" class="form-control" value="<%= month %>" required>
-    </div>
-    <% if (employee.salary_type === 'dihadi') { %>
-      <div class="col-auto">
-        <select name="half" class="form-select">
-          <option value="1" <%= half === 1 ? 'selected' : '' %>>1-15</option>
-          <option value="2" <%= half === 2 ? 'selected' : '' %>>16-end</option>
-        </select>
-      </div>
-    <% } %>
-    <% if (employee.salary_type === 'monthly') { %>
-      <div class="form-check col-auto">
-        <input class="form-check-input" type="checkbox" id="hourly" name="hourly" value="1" <%= hourlyMode ? 'checked' : '' %>>
-        <label class="form-check-label" for="hourly">Hourly view</label>
-      </div>
-    <% } %>
-    <div class="col-auto">
-      <button type="submit" class="btn btn-primary">Go</button>
-    </div>
-  </form>
-  <h5>
-    Month: <%= month %>
-    <% if (employee.salary_type === 'dihadi') { %>
-      - <%= half === 1 ? '1-15' : '16-end' %>
-    <% } %>
-  </h5>
-  <h6>Base Salary: <%= employee.salary %></h6>
-  <table class="table table-bordered">
-    <thead>
-      <tr>
-        <th>Date</th>
-        <th>Status</th>
-        <th>Detailed Status</th>
-        <th>Punch In</th>
-        <th>Punch Out</th>
-        <th>Hours</th>
-        <% if (employee.salary_type === 'monthly') { %>
-          <th>OT Hours</th>
-          <th>UT Hours</th>
-        <% } %>
-        <% if (employee.salary_type === 'dihadi') { %>
-          <th>Lunch Deduction (mins)</th>
-        <% } %>
-        <th>Amount</th>
-      </tr>
-    </thead>
-    <tbody>
-      <% attendance.forEach(a => { %>
-        <tr>
-          <td><%= a.date.toLocaleDateString('en-CA') %></td>
-          <td><%= a.status.charAt(0).toUpperCase() + a.status.slice(1) %></td>
-          <td><%= a.detailed_status || '' %></td>
-          <td><%= a.punch_in || '' %></td>
-          <td><%= a.punch_out || '' %></td>
-          <td><%= a.hours %></td>
-          <% if (employee.salary_type === 'monthly') { %>
-            <td><%= a.overtime %></td>
-            <td><%= a.undertime %></td>
-          <% } %>
-          <% if (employee.salary_type === 'dihadi') { %>
-            <td><%= a.lunch_deduction %></td>
-          <% } %>
-          <td><%= typeof a.amount === 'number' ? a.amount.toFixed(2) : '' %></td>
-        </tr>
-      <% }) %>
-    </tbody>
-  </table>
-  <% if (employee.salary_type === 'monthly') { %>
-    <h6>OT: <%= overtimeFormatted || '00:00' %> | UT: <%= undertimeFormatted || '00:00' %></h6>
-    <% if (hourlyMode) { %>
-      <h6>Total Hours: <%= totalHours || '00:00' %> (Sundays: <%= sundayHours || '00:00' %>) | Hourly Pay: <%= hourlyRate.toFixed(2) %> | Amount: <%= partialAmount ? partialAmount.toFixed(2) : '0.00' %></h6>
-    <% } else { %>
-      <h6>Total Hours: <%= totalHours || '00:00' %> (Sundays: <%= sundayHours || '00:00' %>)</h6>
-    <% } %>
-  <% } %>
-  <% if (employee.salary_type === 'dihadi') { %>
-    <h6>Total Hours: <%= totalHours || '00:00' %> (Sundays: <%= sundayHours || '00:00' %>) | Hourly Pay: <%= hourlyRate.toFixed(2) %> | Amount: <%= partialAmount ? partialAmount.toFixed(2) : '0.00' %></h6>
-  <% } %>
-  <% if (salary) { %>
-    <h6 class="salary-container">
-      Gross: <span class="salary-hidden" data-salary="<%= salary.gross %>">****</span> |
-      Deduction: <span class="salary-hidden" data-salary="<%= salary.deduction %>">****</span> |
-      Net: <span class="salary-hidden" data-salary="<%= salary.net %>">****</span>
-      <button type="button" class="btn btn-outline-secondary btn-sm toggle-salary ms-1"><i class="fas fa-eye"></i></button>
-    </h6>
-    <h6 class="salary-container">
-      Outstanding Advance: <span class="salary-hidden" data-salary="<%= outstanding.toFixed(2) %>">****</span>
-      <button type="button" class="btn btn-outline-secondary btn-sm toggle-salary ms-1"><i class="fas fa-eye"></i></button>
-    </h6>
-    <% if (outstanding > 0) { %>
-      <form action="/employees/<%= employee.id %>/salary/deduct-advance" method="POST" class="row g-2 mb-2">
-        <input type="hidden" name="month" value="<%= salary.month %>">
-        <div class="col-auto">
-          <input type="number" step="0.01" name="amount" class="form-control" max="<%= outstanding %>" placeholder="Deduction amount" required>
-        </div>
-        <div class="col-auto">
-          <button type="submit" class="btn btn-warning">Deduct Advance</button>
-        </div>
-      </form>
-    <% } %>
-  <% } else if (manualSalary) { %>
-    <p>Status: Make manual salary.</p>
-  <% } else { %>
-    <p>No salary record for this month.</p>
-  <% } %>
-</div>
-<script src="https://cdn.jsdelivr.net/npm/bootstrap@5.3.2/dist/js/bootstrap.bundle.min.js"></script>
-<script>
-  const tooltipTriggerList = Array.from(document.querySelectorAll('[data-bs-toggle="tooltip"]'));
-  tooltipTriggerList.forEach(t => new bootstrap.Tooltip(t));
-</script>
-<script src="/public/js/salaryToggle.js"></script>
-</body>
-</html>
