--- views/operatorSalaries.ejs
@@ -1,53 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-  <meta charset="UTF-8">
-  <title>Salary Summary</title>
-  <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.3.2/dist/css/bootstrap.min.css" rel="stylesheet">
-  <style>
-    .toggle-salary { cursor: pointer; }
-  </style>
-</head>
-<body>
-<nav class="navbar navbar-dark bg-dark">
-  <div class="container-fluid">
-    <span class="navbar-brand">Salary Summary</span>
-    <div class="ms-auto">
-      <a href="/logout" class="btn btn-outline-light btn-sm">Logout</a>
-    </div>
-  </div>
-</nav>
-<div class="container my-4">
-  <%- include('partials/flashMessages') %>
-  <table class="table table-bordered">
-    <thead>
-      <tr>
-        <th>Supervisor</th>
-        <th>Active Employees</th>
-        <th>Total Salary</th>
-      </tr>
-    </thead>
-    <tbody>
-      <% summary.forEach(s => { %>
-        <tr>
-          <td><%= s.supervisor_name %></td>
-          <td><%= s.employee_count %></td>
-          <td class="salary-container">
-            <% if (canViewSalary) { %>
-              <span class="salary-hidden" data-salary="<%= s.total_salary %>">****</span>
-              <button type="button" class="btn btn-outline-secondary btn-sm toggle-salary ms-1">
-                <i class="fas fa-eye"></i>
-              </button>
-            <% } else { %>
-              ****
-            <% } %>
-          </td>
-        </tr>
-      <% }) %>
-    </tbody>
-  </table>
-</div>
-<script src="https://cdn.jsdelivr.net/npm/bootstrap@5.3.2/dist/js/bootstrap.bundle.min.js"></script>
-<script src="/public/js/salaryToggle.js"></script>
-</body>
-</html>
