+++ views/supervisorEmployees.ejs
@@ -134,7 +134,6 @@
       <label class="form-label">Salary Type</label>
       <select name="salary_type" class="form-select" required>
         <option value="dihadi">Dihadi</option>
-        <option value="monthly">Monthly</option>
       </select>
     </div>
     <div class="col-md-3">
